Add possibility to override the job result with special comments
When test results are incorporated in other systems and processes the job
result is important and the concept of having a job reviewed with bugref and
labels does not fit all processes. So this is why it can help if the job result
itself can be altered to reflect the actual expectation, for example to mark
false-positives.

Same as for example on o3 with totest-manager we would like to formally
communicate test failures that should be ignored for auto-approval case by
case, e.g. in test failures.

This commit adds parsing for a special comment label "force_result" which can
be used to write in openQA test comments, for example to set a failure to
"softfailed" with a reference to a bug.

The documentation and help text has already been extended in previous commits.

Related progress issue: https://progress.opensuse.org/issues/95479