Display previews for `wait_serial`-results like before 009e9256
* Display previews like before so they are well distinguishable from other
  text results (e.g. produced by `record_info`)
* Does *not* generally revert 009e9256 and in particular the floating
  container will still be used like for any other text result

---

Note that this makes results a little bit more inconsistent again. However,
the problem of not being able to spot normal text results easily within
lots of `wait_serial` results was mentioned by multiple users. I supposed
the main inconsistency which https://progress.opensuse.org/issues/93940 was
about was the expanded view and the missing arrow but not the preview.