Extend test coverage and apply review feedback
Here also i solved a problem which i missed before with the test_log when this is
just one. the problem was in the regex logic which it was always matching only when another git_log
was shown. The regex used to make enumeration of the logs returned by `git log`

Signed-off-by: ybonatakis <ybonatakis@suse.com>