Skip uefi-vars download unless parent has UEFI=1
We already skip the download when cloning a job if `PUBLISH_PFLASH_VARS`
is not present in any parent job to avoid download errors for
non-existing assets. However, it seems some jobs have that variable set
but still don't have `UEFI=1` and are therefore not publishing a
uefi-vars file.

This change is in accordance with the existing condition in
os-autoinst's function `handle_generated_assets`.