clone-job: Fix accidentally skipping asset downloads
When skipping an asset download, the clone-job script skipped all remaining
assets of the same type as well. This is wrong, e.g. if the uefi-vars asset
is skipped before the actual HDD image is considered the actual HDD image
would be skipped as well. With this change only the uefi-vars asset would
be skipped.

Note that this problem was made worse by e0bb6cb73 which added one more
case to skip assets.