Speed up 25-cache-service.t with a shorter worker timeout
The `--dequeue-timeout` option comes from
`Minion::Command::minion::worker` and is used to decide how long the
Minion backend will block when waiting for the next job to be dequeued.
By reducing it from 5 seconds to 1 second we can speed up the worker
shurdown in our tests.

While it is possible to reduce the value further to 0 seconds, and
speed up tests a little more, that also has the negative side effect of
increasing CPU usage very significantly. So it is probably not worth
doing.