t: Provide autoinst-log.txt in full-stack.t also before uploading
There are cases when an openQA job is not yet finished (or about to be
finished with uploading) and optimizing going on. At this time
autoinst-log.txt can not yet be in the testresults directory so we
should look into the pool directory as well.

Related progress issue: https://progress.opensuse.org/issues/102578