t: Add more details to wait_for_result_panel log message
The previous string "result panel shows" sounds like the result panel
would actually show the expected string at that time but we only wait
for that to show.

This commit changes the phrasing and adds the current count and maximum
timeout until we wait. This can help in case we suspect we run into the
timeout or expect that the timeout might be miscalculated.

Related progress issue: https://progress.opensuse.org/issues/102578