t: Fix regression that we fail on too many js errors (48664aa)
This fixes an error introduced with
https://github.com/os-autoinst/openQA/pull/4288
where an additional "if"-expression was introduced looking for two
alternatives "javascript" or "network" for the error source which means
that we never enter the next elsif-sections for such errors.

This commit simply removes the distinction of javascript error sources
as we can be strict enough with our individual regex's looking for
specific errors to skip.

Related progress issue: https://progress.opensuse.org/issues/102578