t: Prevent non-deterministic test step calculation using wait_for_ajax
The test helper function "wait_for_ajax" pretended to be a "test" with a
"pass" instruction when actually it is just a helper that can wait a
non-deterministic amount of time causing the test steps calculation to
be different every time. This means that also on any test failures the
reached test step number provides no useful information. This commit
changes the call to "pass" to "note" instead which still provides output
in verbose mode but does not change the test count.

Related progress issue: https://progress.opensuse.org/issues/102578