CI: Prevent commit message check on master when it is too late
When updating "master" in my personal fork I realized that the git
commit message checker is called and fails due to a commit in master
which by now we can not fix anymore unless we would rewrite the history
in master which we should avoid.

This commit excludes the commit message check to run on pushes to
master.