Avoid Perl warning in `testcasedir` when no root directory is present
The function was returning a non-existing path before in this and other
cases. I suppose it is ok to simply suppress the Perl warning `Use of
uninitialized value in subroutine entry` here by always using the default
project dir (returning a path like `/var/lib/openqa/share/tests/…`).

Note that this warning was showing up in `t/16-utils.t` when
`/var/lib/openqa` does not exist.