Add advanced restart dropdown menu
The additional flags the REST-API offers for restarting can now be used
when restarting a job via the web UI's job details page as well.

This commit changes the restart icon to be a full button instead which
in a dropdown-menu offers additional restart options.

Related progress issue: https://progress.opensuse.org/issues/69979