Move 'dsn' functionality into only place where used
In b1da48eac I removed the "dsn" function and its usage as it looked
like that would not be necessary anymore at all and we did not have any
test coverage for that. This turned out to be a false assumption leading
to a production openQA instance to not start anymore with the error
message

```
openqa-webui-daemon: Can't connect to data source 'HASH(0x55d82c060598)' because I can't work out what driver to use (it doesn't seem to contain a 'dbi:driver:' prefix and the DBI_DRIVER env var is not set)>
```

so the change was quickly reverted.

This commit now moves the 'dsn' functionality to the only place where
it is needed, simplifies the call and adds complete test coverage for
the according method with a separate, new test module
"t/15-shared-plugin-gru.t".

Tests did not cover the 'dsn' function anyway.

Verified by calling

```
env OPENQA_BASEDIR=$PWD/t/data OPENQA_DATABASE=test script/openqa-gru
```

locally.