WIP: use tpage to create the apache templates
Some of the details of the apache .template files in Debian need to be
different, primarily where they are installed and comments about how SSL
is configured.  That being the case we have a patch to do that.

As an alternative to carrying that patch, I thought I might be able to
come up with a template that can combine the creation of both the
non-SSL and SSL templates (hopefully making it worth accepting upstream)
while also making the template and the Makefile a little more flexible
such that I can define a couple of flags in the debian/rules file, and
have it behave as we need for the Debian package.

This introduces a build dependency on Template::Tools::tpage, which I've
not attempted to add to the dependency information, as I'm not sure how
that's done.

If you don't think this is worth having, that's OK -- I'll just continue
with the patch as I already have it.

FYI the things I define for Debian are:

  APACHE_VHOSTS_DIR=sites-available
  APACHE_TEMPLATE_OPTS="--define distri=Debian"

The first ensures that the files are installed into Debian's preferred
/etc/apache2/sites-available (rather than .../vhosts.d seen on openSUSE)
and the second mostly gets rid of the stuff about needing SSL in
APACHE_SERVER_FLAGS (which is not how we do it in Debian)

I note that we also need to enable the `expires` module in apache when
on Debian. I don't know if that's a real difference, or whether perhaps
you also need that in openSUSE and simply forgot to include mention of
that here.