Show "failed modules" with a coloring resembling an failed state
Especially with the new openQA feature of "force_result" it is more
common to see openQA jobs where the result is passed or soft-failed
while there are still failed modules. Right now e.g. in /tests/overview
"failed modules" show up on a green background which could confuse
people thinking of these modules as not failed. We should reconsider
the coloring. This commit changes the styling so that the failed modules
in openQA are rendered with a slightly more alarming design using a
light red background and red border which I made a bit thicker.

Related progress issue: https://progress.opensuse.org/issues/95479