Restart failed chained parents "up the chain"
Normally one can simply restart the failed chained parent directly.
However, when the dependency tree has multiple failing "root" jobs it
is not so easy. If one restarts one of the "root" jobs the other one
would not be restarted as one needed to go "up the chain" to reach the
other "root" job. Simply restarting the other "root" job afterwards is
not going to cut it this would lead to creating a second set of child
jobs which is not wanted (and not even possible unless `clone=0` is
used.

One can still opt-out by using the existing parameter `skip_parents=1`
which is also accessible from the UI.

See https://progress.opensuse.org/issues/95783#note-16 for the concrete
use-case within the investigation script.