t: Fix querying stale reference in fullstack and developer mode test
In these tests it is not possible to use
`$driver->find_element('#info_box .card-body')->get_text` because the
element might be swapped out by JavaScript under the hood after it has been
returned by `find_element` and before the text is queried via `get_text`.

This leads to errors like:

```
[16:04:54] t/full-stack.t .. 35/? getElementText: stale element reference: element is not attached to the page document at /home/squamata/project/t/lib/OpenQA/SeleniumTest.pm:80 at /home/squamata/project/t/lib/OpenQA/SeleniumTest.pm line 83.
…
    Selenium::Remote::WebElement::get_text(Test::Selenium::Remote::WebElement=HASH(0x55903e1b63d8)) called at /home/squamata/project/t/lib/OpenQA/Test/FullstackUtils.pm line 64
    OpenQA::Test::FullstackUtils::find_status_text(Test::Selenium::Chrome=HASH(0x55903dd0fc40)) called at /home/squamata/project/t/lib/OpenQA/Test/FullstackUtils.pm line 82
    OpenQA::Test::FullstackUtils::wait_for_result_panel(Test::Selenium::Chrome=HASH(0x55903dd0fc40), qr(Result: passed)u, "job 6") called at t/full-stack.t line 389
…
```
(see https://progress.opensuse.org/issues/98952#note-54 for details)

Querying the text via JavaScript should fix the race condition.

Note that this is only a problem if the test is not in a final state yet
and these seem to be the only instances of this situation in our test
suite.