Ensure the dependency graph is sorted consistently
So far the order of elements within the dependency tree could be different
depending on the current job. This change makes it consistent by sorting
the edges before passing them to the rendering library. (Sorting nodes does
not help and also does not seem to be required.)

See https://github.com/os-autoinst/openQA/pull/4507#issuecomment-1041556218