Re-try accepting directly chained jobs to avoid skipping whole chain
The worker doesn't retry accepting a job so far. That makes actually sense
because at this point no work has been done anyways and the worker can just
wait for the scheduler to re-assign the job. However, for directly chained
jobs this is not true because the whole chain needed to be restarted in the
case of an error. So we should try a little bit harder - similarly to how
it is already done when the connection is lost during the job execution.

See https://progress.opensuse.org/issues/107746