Validate `groupid` parameter properly to avoid SQL errors in logs
In contrast to the other test overview search parameters it must be an
integer. Otherwise we end up with an SQL error (see
https://progress.opensuse.org/issues/107926).

If the `groupid` is invalid a flash message will be shown about it on the
page. Note that this kind of user feedback is not useful for the other
parameters because we only use the validation to filter out empty values
here. So I'm only covering `groupid` here.