Fix test of openqa-clone-custom-git-refspec
I apparently didn't correctly push the fix for messing a
check of that unit test (as pointed out in
https://github.com/os-autoinst/openQA/pull/4563#pullrequestreview-917187385).

This change restores the test to check for the `TEST` variable as
it was done before (where it was `TEST=my_test\@user/repo#my/branch.*`,
now it is `TEST+=…`). The `BUILD` likely doesn't matter in this
particular check.