worker: Consider job result from web UI when deciding to skip jobs
So far the worker only considered problems on the worker side (e.g. backend
crashed, fatal API failures) and would therefore have executed directly
chained children despite the parent job failed in a job module (or any
problem which is only evaluated by the web UI). This change fixes the
behavior so directly chained children are skipped in that case.

Likely it never was that problematic because skipping jobs was flawed
anyways (so it happened more often than it should have happened anyways).
However, with the previous commit and 49cf1ca0cc this problem would surface
more often.