Fix the force execution mode in fetchneedles script
fetchneedles script support 'force' execution mode to clear
any uncommitted changes before a git update.
This behavior did not always works as expected.
fetchneedles script has some code to call itself with proper
permissions. To do so, most relevant script
arguments are forwarded in the code that self-call the script.
'force' argument was not in the list of forwarder arguments.