Fix minion jobs stats retrieval DoS on login provider
Allow the continuously polled stats to be available on an
unauthenticated route to prevent recurring broken requests to the login
provider if not logged in to prevent unnecessary and harmful recurring
traffic to the login provider service, e.g. id.opensuse.org, which
caused the login provider service to misbehave.

First to reproduce the issue what I did was to start a bare openQA
instance with config `httpsonly=0` in etc/openqa/openqa.ini of my openQA
working copy and keeping everything else in place. Then started `sudo -u
geekotest env OPENQA_BASEDIR=$tmp script/openqa-webui-daemon` with $tmp
being a directory owned by geekotest and an empty subdirectory db to
prevent an error about unable to create a lock file. Then I went to
http://127.0.0.1:9526/, logged in, then to
http://127.0.0.1:9526/minion/jobs in my browser, opened developer tools,
"Network" tab and saw a request every couple of seconds to "stats". I
opened http://127.0.0.1:9526/ in a second tab, logged out and could see
that the requests in the developer tools go to "www.opensuse.org/openid"
instead. With the change applied I could confirm that stats is continued
to be polled from the local server regardless of the authentication
state.

Related progress issue: https://progress.opensuse.org/issues/110533