Prevent job_done_hooks not knowing that jobs will be retried
A common use case for job_done_hooks is to handle failed jobs. However
there can be that case that we already sufficiently handle any such
failures by retrying test cases using the openQA retry feature. So if
test maintainers signal the intent that the right approach to handle
failed jobs is to retry we should ensure that we trigger a retry and
hence put a comment on an openQA job to provide a consistent state
before we trigger any asynchronous job_done_hooks which can evaluate
such information, e.g. as
https://github.com/os-autoinst/scripts/blob/master/openqa-investigate
does to detect if an issue is "known" or "unknown" based on the presence
of a comment.

This effectively fixes a race condition by ensuring synchronous
operations complete before triggering asynchronous jobs.

Related progress issue: https://progress.opensuse.org/issues/110911