Introduce job setting `_TRIGGER_JOB_DONE_HOOK` to disable/enable hook
This setting allows to disable/enable the hook script per job (or test
suite or wherever the setting is specified). It also comes with a generic
`job_done_hook` configuration setting which will be used in case
`_TRIGGER_JOB_DONE_HOOK=1` is specified and no result-specific job done
hook is configured anyways.

See
* https://progress.opensuse.org/issues/110518
* https://progress.opensuse.org/issues/110530