CI: Automatically merge necessary dependency updates
Whenever our development base has updated dependencies we explicitly
update the list of dependencies but old versions are not preserved in
the base OS layer so regardless of if we like the changes or not we need
to accept the changes anyway, in particular version updates. This commit
is extending the mergify configuration to automatically merge pull
requests by the "openqabot" user when all our common checks pass.