Makefile: Ensure "test" includes also helm tests
Running all tests as part of the "test" target is the safer approach to
ensure that no tests are overlooked. Many developers will unfortunately
not have all prerequisites to run all the tests necessary locally but
that was already the case in before. Indidvidual test parts can still be
selected explicitly as well as excluded based on variables. So this
commit adds the helm-chart test to the generic "test" target while
configuring it based on the HELM_TEST variable and same as the
CONTAINER_TEST excluding the helm-chart test when the TESTS variable is
specified to select individual test modules.