CI: Fix execution of compound statements with "retry"
In d7fb5d5 I introduced a compound 'zypper ref && zypper in' call to be
called by tools/retry. At that time I tested the combined zypper
statement itself but did not realize that tools/retry does not execute
the combined "sh" with the command line parameter string.

This commit fixes the problems we observed in circleCI runs on the
master branch trying to create a dependency update pull request.

This change is in line with
https://github.com/okurz/retry/blob/main/retry