Fix sort order of comments on job details page
Comments are accessed via `$job->comments` when rendering them so the
sorting is expected to be ascending by ID via
`__PACKAGE__->has_many(comments => 'OpenQA::Schema::Result::Comments',
'job_id', {order_by => 'id'});` in `Jobs.pm`. However, that's not effective
since we're prefetching the comments here. So it is necessary to specify
the order of the comments when querying the job (where comments are
prefetched).