Fix AJAX/parsing error in needle tables
The date/time picker of Firefox and Chromium passes a timestamp without
seconds when a timestamp on a whole minute is selected. So far this is not
accepted by the controller code leading to an alert box complaining about
an AJAX error. This change fixes the issue by allowing the format on the
controller side. Note that minutes and hours always seem to be present.

This issue has been found due to a test failure when the date/time picker
was randomly on a whole minute (see
https://github.com/os-autoinst/openQA/pull/4791#issuecomment-1234391043 and
further comments).