Use the workers count from Minion 10.25 to make monitoring more reliable
It has been suggested that our monitoring is unreliable because adding
up number of active and inactive workers doesn't quite work right. So we
have added the number of all registered workers upstream to the Minion
stats. The feature will be available with the Minion 10.25 release.

Progress: https://progress.opensuse.org/issues/112898