t: Fix sporadic failure in 25-cache-service
In about 30% of runs t/25-cache-service.t failed for me in the test
"Cache server is not available". Apparently the cache service is still
available in some cases after the "stop" call. Likely the double
"restart" call on the process when starting is creating problems and
also looks wrong and unnecessary. The double restart was introduced in
b95c41051 which was a not so well reviewed pull request which maybe only
introduced the double restart as a "copy-paste-mistake". With the
duplicated restart removed the command

```
count_fail_ratio prove -l t/25-cache-service.t
```

was stable for me over 15 executions so far.