Support skip and broken results in LTP parser
Since now, Perl version of runltp-ng used to generate a specific JSON
file without including TBROK or TCONF results, which have to be
respectively trated as test failure and test skip.

With the introduction of the new Python version of runltp-ng, we are now
generating a JSON file containing TBROK and TCONF tests statuses. For
this reason, we need to allign LTP parser code so jobs will be treated
accordingly.

With this patch jobs statuses will be mapped using the following rules:

- TPASS -> 'ok'
- TFAIL/TBROK -> 'fail'
- TCONF/TSKIP -> 'skip'