Show correct job status in worker job list
Issue: https://progress.opensuse.org/issues/119110

The renderTestName function expected a running to job to have
an undefined or non-existant`result` entry, which was the case for the
test lists on the "All tests" page, but not for the job list on
a worker page.
That's why it was shown with "Done: none".
Additionally it showed a restart button, which made the test
in 15-admin-workers.t restart the first job 99961, although
it is running.

Also the cancel_url variable was never set.