Add an API endpoint to find jobs with certain job settings
Scanning all job settings can be quite costly. On OSD it easily takes
more than 10 seconds.

There are two ways to optimise this, one is to add a trigram gin index
on the value field of the job_settings table. Which on
OSD would take about 20 minutes to create and use 2GB of disk space.
Additionally the pg_trgm extension would need to be activated for all
openQA installations. Which might not be possible to do automatically
with the database migration system we are using.

The other way is to simply limit the number of recent jobs we include in
the search. And that's the approach this commit takes. The default of
20000 jobs has been tested on OSD, and performs quite similar to how a
trigram index would perform. The downside of course is that if 20000
jobs were to be created in quick succession, all older incident data
would be inaccessible via the API. That's why i would personally prefer
the trigram index solution. Maybe we can switch to that in the future.

For more details see also the measurements taken in the progress ticket:
https://progress.opensuse.org/issues/117655