Workaround openQA-single-instance pulling old openQA on Leap 15.4
Using the openQA-bootstrap script on Leap 15.4, without ssh keys in use, causes
it to fail, because it pulls an older version of fetchneedles and openQA where
poo#104794 has not been fixed.

However by explicitly adding openQA to the list of the packages to be installed,
the problems seems to be fixed.

Related ticket: https://progress.opensuse.org/issues/115784