Fix race-condition reading job module details files
A not readable test module result file is handled and gracefully
returned to the web UI and no error in log. The handling should be
atomic, not "check if readable and then read". Instead of trying to read
and then fail we should handle the read attempt as we handle the
exception in line 96 and return {} in case of error.

Related progress issue: https://progress.opensuse.org/issues/121444