Protect second asset size update from possible deadlocks
We are dealing with an update order deadlock here. Imagine two processes
are running an update query for three rows against the same table.
Process 1 updates the rows in order A, B, C. Process 2 updates them in
order C, B, A. Now process 1 holds the lock on A and is waiting for the
lock on C, while process 2 holds the lock on C and is waiting for A. We
have a deadlock that PostgreSQL cannot resolve on its own.

This is a similar case, the random order just happens to come from the
Perl side in form of a `values %assets`. To fix that we ensure a
consistent order based on the asset name. Additionally we also
defensively use a `SELECT...FOR UPDATE` before the `UPDATE`.

Progress: https://progress.opensuse.org/issues/120891