Don't restart scheduled or running chained parents
This is a follow-up to @martchus #4498. The intent of that PR,
as the commit message says, is to "Restart failed chained parents
"up the chain"". However, it actually does more than this: it
also restarts chained parents that are queued or running, because
they also do not have "OK" results. This tweaks the check a bit
so we only restart parents if they have a result which is known
to be not OK. This leaves out NONE, which is the result that
scheduled or running tests have.

In a way we cannot know whether restarting a chained parent that
is still running is the right thing to do, but doing so definitely
causes problems in one real-world case:
https://progress.opensuse.org/projects/openqav3/activity?from=2022-12-19
and from the examples backing the original change here, it seems
we only really intended to add the restart behaviour when the
parent test definitely failed, not when its result was not yet
decided. So on the principle of making minimal changes to the
previous behaviour, it seems reasonable to make this change,
which in fact restores the behaviour from before #4498 for
still-running parent jobs. The cases #4498 intended to address
with complex investigation jobs should still be addressed.

This also updates the comments for `duplicate` to more accurately
reflect current behaviour, and highlight the problem that
ignoring children of a parent job can ignore not only other jobs
in the same parallel group as the child job, but other children of
the same parent that are *not* in the same parallel group as the
child job.

Related progress issue: https://progress.opensuse.org/issues/110458
Related progress issue: https://progress.opensuse.org/issues/112256

Signed-off-by: Adam Williamson <awilliam@redhat.com>