Add postgresql service to requirements
If this unit gets activated, the units listed in Requires= will be
activated as well. If one of the other units fails to activate, and an
ordering dependency After= on the failing unit is set, this unit will
not be started. Besides, with or without specifying After=, this unit
will be stopped (or restarted) if one of the other units is explicitly
stopped (or restarted).

Reference: https://progress.opensuse.org/issues/122578