Fix asset cache error handling to avoid proceeding in case of failures
So far the cache service doesn't treat download errors as such when the
asset is still present locally. However, that's not a good idea because the
local version might be outdated and problems on the server-side might only
surface later (on the next cache cleanup) which makes the investigation
more difficult (see https://progress.opensuse.org/issues/122776).

This change tracks whether the download has actually been succeeded instead
of only checking whether the cached asset exists locally.

Note assets that are locally cached are *not* re-downloaded unless the
asset has changed on the server. However, to check whether the asset has
changed on the server an HTTP request is done that might fail. With this
change this error case is now not ignored anymore.