Avoid another race condition by waiting for the #text element
Similar to the previous race condition. Found by Marius this time, but
i've been unable to replicate it locally. Applying the same
`wait_for_element` to ensure that the comment field is visible should
resolve this as well though. The test code is almost exactly the same.

Progress: https://progress.opensuse.org/issues/121042