Improve appearance of tables
Initially I only noticed that if the last <tr/> in a .table had "display: none"
the bottom border would be missing entirely. The only way to address that was
to redo the way the table border is laid out. This turned out to be shorter
than the original code and fixes other issues like missing borders in dark
mode.

Additionally, vertically center table contents explicitly. This requires
including openqa.scss after bootstrap so that our declarations are preferred.