Render bugrefs as clickable links within labels
So one does not need to add a bugref just for the sake of having a
clickable link. This is especially useful as having a label and a bugref
in one comment has caveats. In particular, a `force_result` will not be
evaluated if the comment is carried over and the carry over itself prevents
hook scripts from running.

With this, the change https://github.com/os-autoinst/scripts/pull/155 could
be reverted.

Related ticket/note: https://progress.opensuse.org/issues/124029#note-4