Avoid caveats of `Archive::Extract` using `bsdtar` instead
We observed extracted assets that are zero-sized on o3. The related jobs
were nevertheless executed. Maybe the error handling of the asset
extraction via `Archive::Extract` is broken (by an internal problem of
`Archive::Extract` and its underlying modules). At least it does not seem
to be an issue on our side (the download itself has proper error handling
and we check for errors reported by `Archive::Extract`).

By simply using `bsdtar` (or `bsdcat`, also in the `bsdtar` package)
instead of relying on Perl modules of questionable qualify we can avoid the
problem.

Note that `bsdcat` will not complain if a file does not appear to be
compressed. It will then just output the file contents as-is. This was
previously an error. I think this is a change in behavior we can live with.

As a plus, with `bsdtar` and `bsdcat` we no longer rely on the extension to
determine the file format. So e.g. an xz compressed file is uncompressed as
expected, even without or a wrong extension. We also gain support for many
more formats like zstd. This change will likely also make the extraction
faster because `Archive::Extract` only used a pure Perl implementation in
the current configuration.

Related ticket/note: https://progress.opensuse.org/issues/123175#note-27