Rename `SCHEDULE_FROM_YAML_FILE` to `SCENARIO_DEFINITIONS_YAML_FILE`
This file does *not* define a concrete list of jobs to be scheduled (a
schedule). It rather defines definitions of projects, machines and job
templates. So it basically defines *all* conceivable test scenarios to be
considered when a new product is scheduled. Only the scenarios from that
file that match the specified product will be actually scheduled.