spec: Fix missing 'sass' during runtime with newer deps
I suspect that a newer AssetPack version needs sass in the runtime
environment or we did not properly update an asset cache in the package
or something. This commit now requires the sass application in
build-time as well as run-time to fix that problem.

Related progress issue: https://progress.opensuse.org/issues/124739