Add openQA status badges
Ticket: https://progress.opensuse.org/issues/124173

The badges can be used to show the current status of an openQA job
eg. within a github PR.

The badges can be accessed for any job via:

/tests/123/badge
/tests/latest/badge

They support the get parameter 'show_build=1' that will
prefix the status with the build number (useful for the /latest/badge).