Improve error handling when automatically restarting a job
The web UI automatically restarts a job when it is set to done under
certain conditions. So far errors when restarting are ignored. This change
moves the restarting into a Minion job. An error will now lead to failing
Minion jobs with the error message assigned as result. This should allow us
to investigate https://progress.opensuse.org/issues/125276. Minion jobs are
also retried a few times before ending up failed which should make it a bit
more robust as well.