Trigger job finalization reliably from livehandler
The livehandler cancels jobs them so it needs to load the Gru
plugin in order to enqueue the Minion job for finalizing job
results (and triggering the hook script).

Note that this problem was only revealed after removing the
`eval` in the function to enqueue the Minion jobs. Before,
this problem was simply ignored.

Note that the scheduler is already covered similarly.