Tweak retry-loop of `openqa-cli`
When configuring exactly one retry, the loop currently does the
following (assuming the request fails and should be retried):
```
1. Request
2. Log "Request failed" and sleep for the configured retry delay
```
So no retry is done despite being logged and sleep being
executed. This is identical to zero retries being configured
except for 2. but should actually lead to two attempts in total.
If more than one retry is configured the problem is the same of
course.

This change alters the behavior to be (and by the way needs two
less lines of code):
```
1. Request
2. Log "Request failed" and sleep for the configured retry delay
3. Do actually one more request
```