Optimize database queries for next/previous
Fetch dependencies and failed modules for all jobs in one go instead
of doing up to 500-1000 * 3 queries.

The duration of that route went down locally from 1.25s to 0.2s for a job with
more than 50 next/previous pages.

As we do this optimization already for the /tests/overview page I put some
code into subroutines.