Fix enabling live stream/log when the backend is not up yet
The current implementation of enabling the live log does nothing
at all when the backend hasn't been started yet. This is a
problem if asset caching is enabled because then the backend
startup is delayed until all assets have been cached and the
frontend likely sends the command for enabling the live log
before the backend has been started.

This change causes the worker to track the number of live log
viewers regardless of whether the backend is running or not.
After the backend has been initialized the current number of
live log viewers is evaluated and handled accordingly (in the
same way as changes in viewership are handled normally).

Related ticket: https://progress.opensuse.org/issues/113318