Improve contrast of Progress bar text
This commit addresses the issue of improvement on the contrast of progress bar text, the change in contrast ensured easy readability by the users especially those with visual impairment. The Contrast of the progress bar was changed from #CACACA to #000000 to ensure better visibility on the bar and accessibility for the users.